update merge lora
